backend/requirements.txt
aiofiles==25.1.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
av==16.1.0
billiard==4.2.4
boto3==1.42.44
botocore==1.42.44
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
ctranslate2==4.7.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
exceptiongroup==1.3.1
fastapi==0.128.6
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.2.0
ftfy==6.3.1
future==1.0.0
grpcio==1.78.0
grpcio-tools==1.59.5
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
kombu==5.6.2
loguru==0.7.3
MarkupSafe==3.0.3
minio==7.2.20
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
onnxruntime==1.23.2
opencv-python-headless==4.11.0.86
packaging==26.0
pillow==12.1.0
platformdirs==4.5.1
portalocker==2.10.1
prompt_toolkit==3.0.52
protobuf==4.25.8
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
qdrant-client==1.10.1
redis==7.1.0
regex==2026.1.15
requests==2.32.5
s3transfer==0.16.0
safetensors==0.7.0
scenedetect==0.6.7.1
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.1.0
triton==3.6.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.6.0
yt-dlp==2026.2.4
